requirements.txt
# Django Portfolio - Requirements

## Production
Django>=5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-dotenv>=1.0
django-environ>=0.11
gunicorn>=22.0
whitenoise>=6.6
django-cors-headers>=4.3
Pillow>=10.3
